\name{adjustPrsWeights}
\title{PRS weight adjustment through final regression}
\usage{
    adjustPrsWeights(snpSelection, gwe, response, 
        covariates = NULL, pcs = FALSE, family = NULL,
        rc = NULL)
}
\arguments{
    \item{snpSelection}{}
    
    \item{gwe}{}
    
    \item{response}{}
    
    \item{covariates}{}
    
    \item{pcs}{}
    
    \item{family}{}
    
    \item{rc}{The fraction of available cores to use for 
    parallel calculations. Default is \code{NULL} for no
    parallelization.}
}
\value{
    
}
\description{
    
}
\details{
    
}
\examples{
    
}
\author{
    Panagiotis Moulos
}

