\name{aggregatePrsMarkers}
\title{Aggregate PRS candidate SNPs from iterative extraction}
\usage{
    aggregatePrsMarkers(gwaList, mode = c("intersect", 
        "union"), qcut = 0.9, assoc = c("auto", "glm",
        "rrblup", "statgen", "snptest", "plink", "lasso"),
        avgfun = c("mean", "median", "weight"), gwe = NULL)
}
\arguments{
    \item{gwaList}{}
    
    \item{mode}{}
    
    \item{qcut}{}
    
    \item{assoc}{}
    
    \item{avgfun}{}
    
    \item{gwe}{}
}
\value{
    
}
\description{
    
}
\details{
    
}
\examples{
    
}
\author{
    Panagiotis Moulos
}

