\name{asap}
\docType{data}
\alias{asap}
\title{The ASAP Data Set}
\description{
  Toy data set to show the capabilities of the PRISMA package.
}
\usage{asap}
\format{A prisma object.}
\references{
Krueger, T., Kraemer, N., Rieck, K. (2011)
ASAP: Automatic Semantics-Aware Analysis of Network Payloads
\emph{Privacy and Security Issues in Data Mining and Machine Learning - International ECML/PKDD Workshop. Lecture Notes in Computer Science 6549}, Springer. 50 - 63
}
\author{
Tammo Krueger <t.krueger@tu-berlin.de>
}
\keyword{datasets}