\name{cvSumFuns}
\alias{cvSummary}
\alias{summarizeCvMetrics}
\title{Functions to split datasets in training and test samples}
\usage{
    cvSummary(M, nsnp, scr = TRUE, out = FALSE)
    
    summarizeCvMetrics(cvm)
    
    plotCvMetrics(cvo, what = c("r2", "rmse", "mae", "pr2","crl"),
        pl = c("pct", "prs"), silent = FALSE)
}
\arguments{
    \item{M}{An output of}
    
    \item{nsnp}{}
    
    \item{scr}{}
    
    \item{out}{}
    
    \item{cvm}{}
    
    \item{cvo}{}
    
    \item{what}{}
    
    \item{pl}{}
    
    \item{silent}{}
}
\value{
    
}
\description{
    
}
\details{
    
}
\examples{
    
}
\author{
    Panagiotis Moulos
}

