\name{ebiQueryFuns}
\alias{getGWASVariants}
\alias{getPGSScores}
\title{Functions to query EBI GWAS and PGS Catalogs}
\usage{
    getGWASVariants(efoId = NULL, efoTrait = NULL, 
        removeUnknownRisk = TRUE, retries = 5)
    
    getPGSScores(pgsId = NULL, efoId = NULL, pubmedId = NULL,
        base = NULL, retries = 5, validateLoc = FALSE)
}
\arguments{
    \item{efoId}{}
    
    \item{efoTrait}{}
    
    \item{pgsId}{}
    
    \item{pubmedId}{}
    
    \item{retries}{}
    
    \item{removeUnknownRisk}{}
    
    \item{base}{}
    
    \item{validateLoc}{}
}
\value{
    Depending on \code{out}
}
\description{
    
}
\details{
    
}
\examples{
    
}
\author{
    Panagiotis Moulos
}

