\name{gweConvertFuns}
\alias{GWASExperiment2gData}
\alias{GWASExperiment2GDS}
\alias{GWASExperimentLiftOver}
\alias{guessHumanGenomeVersion}
\title{Functions to convert GWASExperiment to other classes}
\usage{
    GWASExperiment2gData(obj, covariates = NULL, pcs = FALSE,
        reverse = FALSE)
    
    GWASExperiment2GDS(obj)
}
\arguments{
    \item{obj}{a \code{GWASExperiment} object.}
    
    \item{covariates}{}
    
    \item{pcs}{}
    
    \item{reverse}{}
}
\value{
    
}
\description{
    
}
\details{
    
}
\examples{
    
}
\author{
    Panagiotis Moulos
}

