\name{imputeGWAS}
\title{Genotype imputation for GWASExperiment objects}
\usage{
    imputeGWAS(obj, mode = c("single", "split"),
        failed = c("scrime", "none"), rc = NULL)
}
\arguments{
    \item{obj}{}
    
    \item{mode}{}
    
    \item{failed}{}
    
    \item{selection}{}
    
    \item{genome}{}
    
    \item{alleleOrder}{}
    
    \item{gdsfile}{}
    
    \item{rc}{The fraction of available cores to use for 
    parallel calculations. Default is \code{NULL} for no
    parallelization.}
}
\value{
    
}
\description{
    
}
\details{
    
}
\examples{
    
}
\author{
    Panagiotis Moulos
}

