\docType{data}
\name{pgd}
\alias{pgd}
\title{A toy GWASExperiment dataset with 25 SNPs}
\format{A small \code{GWASExperiment} for demo purposes.}
\source{
    Kalafati et al. (https://doi.org/10.1016/j.nut.2018.10.032)
}
\description{
    This data set contains genotypes for 25 SNPs from the 
    case/control study from the specified source. It is used
    in examples and tests of the prisma package with respect
    to the external PGS analysis.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}

