\name{PRISMA-package}
\alias{PRISMA-package}
\alias{PRISMA}
\docType{package}
\title{
\packageTitle{PRISMA}
}
\description{
\packageDescription{PRISMA}
}
\details{
\packageDESCRIPTION{PRISMA}
\packageIndices{PRISMA}
}
\author{
\packageAuthor{PRISMA}

Maintainer: \packageMaintainer{PRISMA}
}
\references{
Krueger, T., Gascon, H., Kraemer, N., Rieck, K. (2012)
Learning Stateful Models for Network Honeypots
\emph{5th ACM Workshop on Artificial Intelligence and Security (AISEC 2012)}, accepted

Krueger, T., Kraemer, N., Rieck, K. (2011)
ASAP: Automatic Semantics-Aware Analysis of Network Payloads
\emph{Privacy and Security Issues in Data Mining and Machine Learning - International ECML/PKDD Workshop. Lecture Notes in Computer Science 6549}, Springer. 50 - 63
}
\keyword{ package }
\examples{
# please see the vingette for examples
}
