\docType{data}
\name{prismaOut}
\alias{prismaOut}
\title{A sample output from the prisma pipeline}
\format{a \code{prisma} function output \code{list}.}
\source{
    Kalafati et al. (https://doi.org/10.1016/j.nut.2018.10.032)
}
\description{
    This is an output from the \code{prisma} function produced
    from the toy dataset included with the package. It is used
    for speed purposes in \code{prisma} downstream functions.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}
