\name{prsRegressionMetrics}
\title{Calculate PRS evaluation metrics}
\usage{
    prsRegressionMetrics(snpSelection, gwe, response, 
        covariates = NULL, pcs = FALSE, step = 10, family = NULL,
        rc = NULL, ...)
}
\arguments{
    \item{snpSelection}{}
    
    \item{gwe}{}
    
    \item{response}{}
    
    \item{covariates}{}
    
    \item{pcs}{}
    
    \item{step}{}
    
    \item{family}{}
    
    \item{rc}{The fraction of available cores to use for 
    parallel calculations. Default is \code{NULL} for no
    parallelization.}
}
\value{
    
}
\description{
    
}
\details{
    
}
\examples{
    
}
\author{
    Panagiotis Moulos
}

