\name{rsLocsFromEnsembl}
\title{Fill up missing SNP locations using biomaRt}
\usage{
    rsLocsFromEnsembl(rs, gv = c("hg38", "hg19"), canonical = TRUE)
}
\arguments{
    \item{rs}{a character vector of SNP (rs) ids.}
    
    \item{gv}{human genome version (\code{"hg38"}, \code{"hg19"}).}
    
    \item{canonical}{check only canonical chromosomes? Defaults
    to \code{TRUE}.}
}
\value{
    
}
\description{
    
}
\details{
    
}
\examples{
    
}
\author{
    Panagiotis Moulos
}

