\name{thesis}
\docType{data}
\alias{thesis}
\title{The Thesis Data Set}
\description{
  The 15 sections of a thesis (see references) as a tm-corpus.
}
\usage{thesis}
\format{A tm-corpus.}
\references{
Tammo Krueger. \emph{Probabilistic Methods for Network Security. From Analysis to Response.} PhD thesis,
TU Berlin, 2013. \url{http://opus.kobv.de/tuberlin/volltexte/2013/3881/}
}
\author{
Tammo Krueger <tammokrueger@googlemail.com>
}
\keyword{datasets}