\docType{data}
\name{toy}
\alias{toy}
\title{A toy GWASExperiment dataset with 2 chromosomes}
\format{A small \code{GWASExperiment} for demo purposes.}
\source{
    Kalafati et al. (https://doi.org/10.1016/j.nut.2018.10.032)
}
\description{
    This data set contains genotypes for chromosomes 18 and 21
    from the case/control study from the specified source. It
    is used in examples and tests of the prisma package.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}
